fix: move home page to customers
